refactor: move SearchContext and SearchStrategy to services
